+++ ChangeLog.md
@@ -9,6 +9,7 @@
 - BUGFIX: #4198 When copying/moving table with name-colliding index (or trigger) to another database, the name conflict resolution now actually works.
 - BUGFIX: #3541 Standardized commit/rollback hotkeys for: functions editor, extensions editor, collations editor, snippets editor, table window structure tab, view window query tab.
 - BUGFIX: #4607 Fixed bug in Query Executor, that caused selecting ROWID from a View, without expanding the view in the first place, thus causing smart execution to fail.
+- BUGFIX: #4614 Fixed opening object links in SQL editor to be case-insensitive.
 - BUGFIX: Fixed refreshing invalid database state if edited & fixed its connection options.
 
 ### 3.4.1